fix(KEN-638): Address PR review - one owner for the move to the trash
The absence guard read every stat failure as absence. An unreadable path
(EACCES on an ancestor, ELOOP, ESTALE) reported the removal applied while
the files stayed installed, and the plan's trailing WriteLock then dropped
the record. It now splits NotFound from unreadable the way pi_hooks_move's
look does, through one symlink_metadata that a dangling link still passes.

relocate was copy_tree's loop body written out again, so how one entry is
reproduced elsewhere had two owners and the duplicate was the symlink arm
the fix rests on. It moves to crate::fs as copy_any, which copy_tree's
loop now calls, under a move_any that owns the rename and its fallback.
pi_ext's own trash, which carried both defects with no symlink arm at all,
becomes one call to it. Its free-name loop asks about a link too, as
unique_in already does.

move_any carries rename's refusal into whatever the second attempt fails
with. Discarded, a refusal a copy does not share read as nothing worse
than "could not remove the original".

Coverage: the route through execute was untested, so the fix is now driven
end to end across a real filesystem boundary (/dev/shm, skipped where that
is not a second mount) with a dangling link, and beside it an unreadable
path that must stop the apply and a dangling link that must reach the
trash rather than be skipped. Each dies to its own mutation.

TrashGuard's doc gave the second op finding nothing as the reason for the
trashed set, which this change makes false; it now gives the reason that
survives.
