+++ CHANGELOG.md
@@ -64,8 +64,8 @@ an outside contributor.
   other row is untouched; rows a pre-rename tool wrote are removed by hand once.
 
 - Removing a skill whose harness copies are only partly present now finishes.
-  A link whose target was already gone failed the move to the trash with "No
-  such file or directory" and rolled the whole removal back, on every retry.
+  A copy that was already gone, or a link whose target was, failed the move to
+  the trash and rolled the whole removal back, on every retry.
 
 - The `task-completed-check` hook counts untracked files as changes, and blocks
   on any nonzero clippy exit or a git that cannot say what changed. It passed
