+++ crates/core/src/apply/op.rs
@@ -154,9 +154,18 @@ impl Op {
                 // half-present with no way forward. Nothing here is
                 // nothing to protect either, so the precondition, which
                 // binds the op to bytes it may take, has nothing to bind
-                // to. A path still holding something is proven as before.
-                if !path.exists() && !path.is_symlink() {
-                    return Ok(());
+                // to.
+                //
+                // Absence proven by the stat, never inferred from its
+                // failure: an unreadable path is one this op knows nothing
+                // about, and calling it removed would take the item off the
+                // books while its files stay installed and still load. Asked
+                // without following a link, so a link whose target is gone is
+                // still here and still proven.
+                match fs::symlink_metadata(path) {
+                    Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(()),
+                    Err(error) => return Err(CoreError::io(path, error)),
+                    Ok(_) => {}
                 }
                 pre.check(path)?;
                 let trash = env.trash_dir();
@@ -165,11 +174,7 @@ impl Op {
                     .file_name()
                     .map(|n| n.to_string_lossy().into_owned())
                     .unwrap_or_else(|| "item".to_owned());
-                let dest = unique_in(&trash, &base);
-                if fs::rename(path, &dest).is_err() {
-                    relocate(path, &dest)?;
-                }
-                Ok(())
+                crate::fs::move_any(path, &unique_in(&trash, &base))
             }
             Op::EditFile { path, edits, pre } => {
                 pre.check(path)?;
@@ -219,30 +224,6 @@ impl Op {
             } => git_config_swap(file, key, expected.as_deref(), value.as_deref()),
         }
     }
-}
-
-/// rename(2) refused the move, so the artifact is reproduced in the trash
-/// and taken away here. The trash on another filesystem than the artifact
-/// is the everyday shape, not a rarity: a project under its own mount is
-/// removed into a trash under the home directory.
-///
-/// A link is remade as a link, and that question is asked first, because
-/// both of the other two read through one. `is_dir` on a link to a
-/// directory sends the shared tree it points at into the trash while only
-/// the link is taken away, and `copy` on a link whose target is gone —
-/// which is what a half-present installation looks like — fails with the
-/// target's ENOENT under the link's name, aborting the removal that was
-/// called to clear it.
-fn relocate(path: &Path, dest: &Path) -> Result<()> {
-    if path.is_symlink() {
-        let target = fs::read_link(path).map_err(|e| CoreError::io(path, e))?;
-        crate::fs::make_symlink(&target, dest)?;
-    } else if path.is_dir() {
-        crate::fs::copy_tree(path, dest)?;
-    } else {
-        fs::copy(path, dest).map_err(|e| CoreError::io(path, e))?;
-    }
-    crate::fs::remove_any(path)
 }
 
 fn write_tree(root: &Path, files: &[(PathBuf, Vec<u8>)], pre: &Pre) -> Result<()> {
@@ -365,6 +346,3 @@ impl Plan {
         self.ops.is_empty()
     }
 }
-
-#[cfg(test)]
-mod tests;
